@@ -16,19 +16,19 @@
     "format": "pnpify run prettier --plugin prettier-plugin-svelte --write src"
   },
   "dependencies": {
-    "@astrojs/svelte": "^5.7.1",
-    "@codemirror/commands": "^6.6.2",
+    "@astrojs/svelte": "^5.7.2",
+    "@codemirror/commands": "^6.7.1",
     "@codemirror/lang-html": "6.4.9",
-    "@codemirror/language": "6.10.2",
+    "@codemirror/language": "6.10.3",
     "@codemirror/state": "6.4.1",
-    "@codemirror/view": "^6.33.0",
-    "@hotwired/turbo": "8.0.10",
-    "@iconify-json/mdi": "^1.2.0",
-    "@lezer/common": "^1.2.1",
+    "@codemirror/view": "^6.34.1",
+    "@hotwired/turbo": "8.0.12",
+    "@iconify-json/mdi": "^1.2.1",
+    "@lezer/common": "^1.2.3",
     "@lezer/highlight": "^1.2.1",
     "@rollup/plugin-yaml": "^4.1.2",
     "@vite-pwa/astro": "^0.4.3",
-    "astro": "^4.16.1",
+    "astro": "^4.16.7",
     "astro-icon": "1.1.1",
     "comlink": "^4.4.1",
     "lodash": "4.17.21",
@@ -37,11 +37,11 @@
     "postcss-import": "16.1.0",
     "postcss-load-config": "^6.0.1",
     "postcss-loader": "8.1.1",
-    "postcss-preset-env": "9.6.0",
+    "postcss-preset-env": "10.0.8",
     "postcss-reporter": "7.1.0",
     "rehype-external-links": "^3.0.0",
-    "svelte": "4.2.19",
-    "vite": "^5.4.6",
+    "svelte": "5.1.3",
+    "vite": "^5.4.10",
     "vite-plugin-pwa": "^0.20.5",
     "workbox-background-sync": "^7.1.0",
     "workbox-core": "^7.1.0",
@@ -53,15 +53,15 @@
     "workbox-window": "7.1.0"
   },
   "devDependencies": {
-    "@yarnpkg/pnpify": "^4.1.2",
+    "@yarnpkg/pnpify": "^4.1.3",
     "astro-eslint-parser": "^1.0.3",
     "eslint": "8.57.0",
     "eslint-plugin-astro": "^0.33.1",
-    "eslint-plugin-svelte3": "^4.0.0",
+    "eslint-plugin-svelte": "^2.46.0",
     "prettier": "^3.3.3",
-    "prettier-plugin-svelte": "^3.2.6",
+    "prettier-plugin-svelte": "^3.2.7",
     "shikiji-core": "^0.10.2",
-    "typescript": "^5.6.2"
+    "typescript": "^5.6.3"
   },
   "packageManager": "yarn@4.0.2"
 }
