@@ -1,5 +1,3 @@
-<svelte:options immutable="{true}" />
-
 <script>
   import { onMount, onDestroy, getContext } from 'svelte'
   import { EditorState, EditorSelection } from '@codemirror/state'
@@ -34,10 +32,10 @@
 
   const TOOLTIP_SHIFT_PX = 20
 
-  let editorElement
-  let editorView = null
-  let tooltipElement = null
-  let tooltipTextElement = null
+  let editorElement = $state(null)
+  let editorView = $state(null)
+  let tooltipElement = $state(null)
+  let tooltipTextElement = $state(null)
 
   const { getWebWorker } = getContext('ww')
 
@@ -257,7 +255,8 @@
       .finally(() => reportLoading.set(false))
   }
 
-  const onSubmitHtml = () => {
+  const onSubmitHtml = (e) => {
+    e.preventDefault()
     reportError.set(null)
     reportLoading.set(true)
 
@@ -336,16 +335,14 @@
 </script>
 
 <div class="editor-area">
-  <div class="editor-area-edit" bind:this="{editorElement}"></div>
+  <div class="editor-area-edit" bind:this={editorElement}></div>
 </div>
 <div class="editor-footer">
-  <button class="editor-area-btn" on:click|preventDefault="{onSubmitHtml}"
-    >Check email HTML and CSS</button
-  >
+  <button class="editor-area-btn" onclick={onSubmitHtml}>Check email HTML and CSS</button>
 </div>
 <!-- tooltip -->
-<div class="editor-tooltip" bind:this="{tooltipElement}">
-  <div class="editor-tooltip-message" bind:this="{tooltipTextElement}"></div>
+<div class="editor-tooltip" bind:this={tooltipElement}>
+  <div class="editor-tooltip-message" bind:this={tooltipTextElement}></div>
 </div>
 
 <style>
