@@ -1,5 +1,3 @@
-<svelte:options immutable="{true}" />
-
 <script>
   import { onMount } from 'svelte'
   import { report, linesAndSelectors } from '@stores/report'
@@ -61,34 +59,34 @@
 <ul class="report-list">
   {#if $report[REPORT_HTML5_DOCTYPE.key] && $report[REPORT_HTML5_DOCTYPE.key].lines.length > 0}
     <ReportItemComponent
-      reportInfo="{REPORT_HTML5_DOCTYPE}"
-      itemName="{''}"
-      itemVal="{''}"
-      elementID="{genElementID([REPORT_HTML5_DOCTYPE, '', ''])}"
-      report="{$report[REPORT_HTML5_DOCTYPE.key]}"
-      handleLineClick="{handleLineClick}"
+      reportInfo={REPORT_HTML5_DOCTYPE}
+      itemName={''}
+      itemVal={''}
+      elementID={genElementID([REPORT_HTML5_DOCTYPE, '', ''])}
+      report={$report[REPORT_HTML5_DOCTYPE.key]}
+      handleLineClick={handleLineClick}
     />
   {/if}
 
   {#if $report[REPORT_CSS_IMPORTANT.key] && $report[REPORT_CSS_IMPORTANT.key].lines.length > 0}
     <ReportItemComponent
-      reportInfo="{REPORT_CSS_IMPORTANT}"
-      itemName="{''}"
-      itemVal="{''}"
-      elementID="{genElementID([REPORT_CSS_IMPORTANT, '', ''])}"
-      report="{$report[REPORT_CSS_IMPORTANT.key]}"
-      handleLineClick="{handleLineClick}"
+      reportInfo={REPORT_CSS_IMPORTANT}
+      itemName={''}
+      itemVal={''}
+      elementID={genElementID([REPORT_CSS_IMPORTANT, '', ''])}
+      report={$report[REPORT_CSS_IMPORTANT.key]}
+      handleLineClick={handleLineClick}
     />
   {/if}
 
   {#if $report[REPORT_CSS_VARIABLES.key] && $report[REPORT_CSS_VARIABLES.key].lines.length > 0}
     <ReportItemComponent
-      reportInfo="{REPORT_CSS_VARIABLES}"
-      itemName="{''}"
-      itemVal="{''}"
-      elementID="{genElementID([REPORT_CSS_VARIABLES, '', ''])}"
-      report="{$report[REPORT_CSS_VARIABLES.key]}"
-      handleLineClick="{handleLineClick}"
+      reportInfo={REPORT_CSS_VARIABLES}
+      itemName={''}
+      itemVal={''}
+      elementID={genElementID([REPORT_CSS_VARIABLES, '', ''])}
+      report={$report[REPORT_CSS_VARIABLES.key]}
+      handleLineClick={handleLineClick}
     />
   {/if}
 
@@ -97,12 +95,12 @@
       {#each Object.keys($report[reportInfo.key]).sort() as itemName (itemName)}
         {#each Object.keys($report[reportInfo.key][itemName]).sort() as itemVal (itemVal)}
           <ReportItemComponent
-            reportInfo="{reportInfo}"
-            itemName="{itemName}"
-            itemVal="{itemVal}"
-            elementID="{genElementID([reportInfo, itemName, itemVal])}"
-            report="{$report[reportInfo.key][itemName][itemVal]}"
-            handleLineClick="{handleLineClick}"
+            reportInfo={reportInfo}
+            itemName={itemName}
+            itemVal={itemVal}
+            elementID={genElementID([reportInfo, itemName, itemVal])}
+            report={$report[reportInfo.key][itemName][itemVal]}
+            handleLineClick={handleLineClick}
           />
         {/each}
       {/each}
@@ -113,12 +111,12 @@
     {#if $report[reportInfo.key]}
       {#each Object.keys($report[reportInfo.key]).sort() as itemName (itemName)}
         <ReportItemComponent
-          reportInfo="{reportInfo}"
-          itemName="{itemName}"
-          itemVal="{''}"
-          elementID="{genElementID([reportInfo, itemName, ''])}"
-          report="{$report[reportInfo.key][itemName]}"
-          handleLineClick="{handleLineClick}"
+          reportInfo={reportInfo}
+          itemName={itemName}
+          itemVal={''}
+          elementID={genElementID([reportInfo, itemName, ''])}
+          report={$report[reportInfo.key][itemName]}
+          handleLineClick={handleLineClick}
         />
       {/each}
     {/if}
